Description: Simple module which checks if the currently url has an alias set and does a 301 redirect to it does.
Reason: Once enabled, an alias provides a nice URL for a node. However it doesn't remove the old URL (eh node/1234). This is a problem as you now have two URL's representing precisely the same content. You're getting into dangerous territory for duplicate pages which can get you sandboxed by the likes of Google!

Solution: This module uses hook_init to interrupt the page load and do an alias lookup. If one exists, redirect off to iit - otherwise the page load may continue uninterrupted. An example of this in use is on the site it was developed for:
http://www.sportbusiness.com/node/160559
redirects to:
http://www.sportbusiness.com/news/160559/lagardere-sets-up-sports-division